refactor: separate executor bean creation from async configurer
